app/server/requirements.txt
coloredlogs
pycryptodome
revChatGPT